feat: rename param package to data
